Add files via upload
One of the most critical issues that the finance sector faces is fraud. The fraud impacts the bottom line of a financial institution. It is estimated that a typical financial institution loses  of its revenue to fraud. If we apply this estimate to the Gross World Product of $79.6, the global loss during 2017 was $4 trillion (more than the GDP of India)
One of the most critical issues that the finance sector faces is fraud. The fraud impacts the bottom line of a financial institution. It is estimated that a typical financial institution loses  of its revenue to fraud. If we apply this estimate to the Gross World Product of $79.6, the global loss during 2017 was $4 trillion (more than the GDP of India)

Problem Statement: 
The goal of the project is to detect whether a transaction is a normal payment or a fraud.

Python package
NumPy 
pandas
matplotlib
SciPy
Sckit-learn(or sklearn)
statsmodel
seaborn